fixed code

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate (double height, double mass){
+    public double calculate(double height, double mass) {
         double index;
-        return (int) (index = mass / (height * height));
+        return (index = mass / (height * height));
     }
 }
